requirements.txt
discord.py
python-dotenv
aiosqlite
aiohttp
pyyaml
requests